add RecipeRevolutionDbContext , connectionString and first Migration

+++ RecipeRevolution.Domain/Entities/NutritionalValues.cs
@@ -9,10 +9,10 @@ namespace RecipeRevolution.Domain.Entities
     public class NutritionalValues
     {
         public int NutritionalValuesId { get; set; }
-        public decimal Calories { get; set; }
-        public decimal Protein { get; set; }
-        public decimal Carbohydrates { get; set; }
-        public decimal Fat { get; set; }
+        public double Calories { get; set; }
+        public double Protein { get; set; }
+        public double Carbohydrates { get; set; }
+        public double Fat { get; set; }
 
         public int RecipeId { get; set; }
         public Recipe Recipe { get; set; }
